don't overwrite non-updated data
Most of the updated properties are stored in lists. Prior to this commit, these lists were replaced with a single, new entry. While this might be OK, I did notice that there was existing data in the destination that would have been deleted.

Also, note that the Organization list isn't preserved. There is no established 'type' into which the sync should store data.